+++ src/Domain/Model/Order/Payment/PaymentStateReader.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Thinktomorrow\Trader\Domain\Model\Order\Payment;
+
+interface PaymentStateReader
+{
+    public function getPaymentState(): string;
+}