+++ src/Domain/Model/Order/Shipping/ShippingStateReader.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Thinktomorrow\Trader\Domain\Model\Order\Shipping;
+
+interface ShippingStateReader
+{
+    public function getShippingState(): string;
+}