+++ src/Domain/Model/Order/State/OrderStateMachine.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Thinktomorrow\Trader\Domain\Model\Order\State;
 
 use Assert\Assertion;
-use Thinktomorrow\Trader\Application\Order\MerchantOrder\MerchantOrder;
 use Thinktomorrow\Trader\Domain\Common\State\AbstractStateMachine;
 use Thinktomorrow\Trader\Domain\Common\State\State;
 use Thinktomorrow\Trader\Domain\Model\Order\Order;
@@ -14,7 +13,7 @@ class OrderStateMachine extends AbstractStateMachine
 {
     protected function getState($model): State
     {
-        if ($model instanceof MerchantOrder) {
+        if ($model instanceof OrderStateReader) {
             // Get class of state so we can create the state
             $firstState = reset($this->states);