+++ src/Domain/Model/Order/State/OrderStateReader.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Thinktomorrow\Trader\Domain\Model\Order\State;
+
+interface OrderStateReader
+{
+    public function getState(): string;
+}